refactor: implement pure compile-time safety checks with zero runtime overhead
- Remove all runtime debugMode variables and checks
- Implement build-tag driven safety checks (unsafe_no_check tag)
- Create separate _dev.go and _prod.go implementations
- Add testingSkipSafetyCheck for test control in development builds
- Split tests requiring safety checks into _dev.go test files
- Fix all test logic for proper safety check toggling
- Ensure zero overhead in production builds with no-op implementations

This change follows CodeRabbit's suggestion to achieve true zero-cost
abstractions in production builds while maintaining safety during development.
