+++ pkg/kernel/kbuffer/BUILD.bazel
@@ -71,6 +71,7 @@ go_test(
         "global_test.go",
         "goroutine_check_test.go",
         "interface_test.go",
+        "kbuffer_bench_test.go",
         "safe_buffer_test.go",
         "safe_sharded_test.go",
         "unsafe_buffer_test.go",
