+++ pkg/kernel/kbuffer/goroutine_check_dev.go
@@ -5,10 +5,9 @@ package kbuffer
 
 // checkSafety performs goroutine safety checks in development builds
 // This is only compiled when unsafe_no_check tag is NOT present
-//
-//go:inline
 func (g *goroutineChecker) checkSafety() {
-	if debugMode {
+	// Allow tests to skip safety checks
+	if !testingSkipSafetyCheck {
 		g.checkGoroutineSafety()
 	}
 }
