+++ pkg/kernel/kbuffer/goroutine_check_prod.go
@@ -3,16 +3,10 @@
 
 package kbuffer
 
-// init sets debugMode to false for production builds
-func init() {
-	debugMode = false
-}
-
 // checkSafety is a no-op in production builds (unsafe_no_check)
 // This completely eliminates any runtime overhead
 //
 //go:nosplit
-//go:inline
 func (g *goroutineChecker) checkSafety() {
 	// No-op: safety checks are disabled in production
 	// The compiler will inline this empty function
