+++ pkg/kernel/kbuffer/goroutine_check_prod_test.go
@@ -11,7 +11,7 @@ import (
 func TestProductionBuildTag(t *testing.T) {
 	// Create an unsafe buffer which uses goroutine checking internally
 	buf := NewUnsafeBuffer(1024)
-	
+
 	// Perform multiple writes that would trigger safety checks in dev mode
 	data := []byte("test")
 	for i := 0; i < 10; i++ {
@@ -23,7 +23,7 @@ func TestProductionBuildTag(t *testing.T) {
 			t.Fatalf("Write returned %d, want %d", n, len(data))
 		}
 	}
-	
+
 	// In production mode, safety checks should be disabled
 	// The buffer should function normally without any panics
 	// and without incrementing internal counters
@@ -52,11 +52,11 @@ func TestProductionInit(t *testing.T) {
 	// Create multiple unsafe buffers and verify they work without safety checks
 	buf1 := NewUnsafeBuffer(256)
 	buf2 := NewUnsafeBuffer(256)
-	
+
 	// Write to both buffers
 	buf1.Write([]byte("buffer1"))
 	buf2.Write([]byte("buffer2"))
-	
+
 	// In production, concurrent access from different goroutines should not panic
 	// (though it's still unsafe and may corrupt data)
 	done := make(chan bool)
@@ -70,9 +70,9 @@ func TestProductionInit(t *testing.T) {
 		}()
 		buf1.Write([]byte("goroutine"))
 	}()
-	
+
 	<-done
-	
+
 	// Buffers should still be functional
 	if buf1.Len() == 0 || buf2.Len() == 0 {
 		t.Error("Buffers should contain data")
